pyproject.toml
[project]
name = "baml-kuzu-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    
    "fastapi>=0.115.13",
    "openai>=1.3.5",
    "ipykernel>=6.29.5",
    "kuzu>=0.10.0",
    "matplotlib>=3.10.3",
    "polars>=1.23.0",
    "pyarrow>=19.0.1",
    "pydantic>=2.10.6",
    "pytest>=8.4.1",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.42.2",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
